Release of X-CUBE-AZRTOS-H7 v2.0.0

+++ Middlewares/ST/levelx/common/drivers/lx_stm32_nand_simulator_driver.c
@@ -237,7 +237,7 @@ UINT    i, j;
     pointer =  (ULONG *) &nand_memory_area[0];
 
     /* Loop to erase block.  */
-    words =  sizeof(nand_memory_area)/sizeof(ULONG);
+    words =  (ULONG)(sizeof(nand_memory_area))/sizeof(ULONG);
     while (words--)
     {
